Fix macOS build by specifying minimum SDK version target (#60)
This fixes the [currently failing
build](https://github.com/calliphlox/cpx-python/actions/runs/4056333092/jobs/6980692533#step:3:273)
for the macOS CI runner. The main fix here is to explicitly specify
10.15 as the minimum SDK version to provide support for some of the
C++20 features we're using. That fix occurs in `build.rs` - there may be
better places to put it (e.g. `pyproject.toml`, `cargo.toml`), but we
haven't found them yet. Previously, the rust/python build chain was
using the default defined by `rustc` which is 10.7.

Unfortunately, I also needed to hard-code the path to the blosc static
libraries as with Windows, which could be annoying to maintain.