Write zarr according to ome-ngff spec (#61)
This makes the corresponding changes to the Python API and tests to
support ome-ngff and pixel sizes.

Pixel scale/size is specified as a 2-tuple of doubles where the values
are in microns. We use the [Python reference
reader](https://github.com/ome/ome-zarr-py) to test that the written
data is read back in correctly.

See https://github.com/calliphlox/cpx/pull/110 for the required changes
to cpx.