Update TriggerEdge enum members to match C headers (#95)
This modifies the Rust definition and therefore Python definition of the
`TriggerEdge` enum so that it defines the same members as [the
corresponding C
enum](https://github.com/acquire-project/acquire-core-libs/blob/40e71b2af97f8827be5d88352414365193145ed3/src/acquire-device-properties/device/props/components.h#L36)
up to the `TriggerEdgeCount` member.

Technically, this is a breaking change to the Python API because it
removes the `NotApplicable` member, which is beyond the count member in
the C definition. We could consider keeping the `NotApplicable` member
to avoid that breaking change.